@@ -1,6 +1,10 @@
-using System;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Typo.Models
+namespace Typo.Model.Models
 {
     public class ErrorViewModel
     {
