@@ -3,16 +3,18 @@ using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Linq;
 using System.Threading.Tasks;
+using Logic;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
-using Typo.Models;
+using Typo.Controllers;
+using Typo.Model.Models;
+
 
 namespace Typo.Controllers
 {
     public class AccountController : Controller
     {
         readonly string _connectionString;
-
         public AccountController(IConfiguration config)
         {
             _connectionString = config.GetSection("Connection")["MSSQL"];
@@ -30,7 +32,6 @@ namespace Typo.Controllers
         [HttpPost]
         public IActionResult Register(Account accounts)
         {
-
             var query = "INSERT INTO Account(username, password) VALUES('{0}', '{1}')";
             var queryFull = string.Format(query, accounts.username, accounts.password);
             var sc = new SqlConnection(_connectionString);
