@@ -1,4 +1,4 @@
-﻿@model Account
+﻿@model Typo.Model.Models.Account
 
 @section Scripts {
     @{ await Html.RenderPartialAsync("_ValidationScriptsPartial"); }
@@ -14,13 +14,13 @@
                 <form method="post">
                     <div class="well">
                         <div class="form-group">
-                            @* <h3 style="line-height: 20%;"><i class="fa fa-home fa-1x" style="line-height: 6%; color: #339966"></i> FirstName:</h3> *@
+                             <h3 style="line-height: 20%;"><i class="fa fa-home fa-1x" style="line-height: 6%; color: #339966"></i> FirstName:</h3> 
                             <label asp-for="username"></label>
                             <input asp-for="username" class="form-control" placeholder="Fill Username" data-val="true" />
                             <span asp-validation-for="username" class="text-danger"></span>
                         </div>
                         <div class="form-group">
-                            @* <h3 style="line-height: 20%;"><i class="fa fa-envelope fa-1x" style="line-height: 6%; color: #339966"></i> LastName:</h3> *@
+                             <h3 style="line-height: 20%;"><i class="fa fa-envelope fa-1x" style="line-height: 6%; color: #339966"></i> LastName:</h3> 
                             <label asp-for="password"></label>
                             <input asp-for="password" type="password" class="form-control" placeholder="Fill Password" />
                             <span asp-validation-for="password" class="text-danger"></span>
@@ -29,9 +29,8 @@
                         <button onclick="location.href = '@Url.Action("Register", "Account")'" class="btn btn-default" type="button">Register</button>
                     </div>
                 </form>
-                
                 @ViewData["Message"]
             </div>
         </div>
     </div>
-</div>
+</div> 
