@@ -1,4 +1,4 @@
-﻿@model Account
+﻿@model Typo.Model.Models.Account
 @section Scripts {
     @{ await Html.RenderPartialAsync("_ValidationScriptsPartial"); }
 }
