@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model Typo.Model.Models.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
