@@ -1,4 +1,5 @@
-﻿<!DOCTYPE html>
+﻿@using Typo.Model.Models
+<!DOCTYPE html>
 <html>
 <head>
     <meta charset="utf-8" />
